requirements.txt
alembic==1.12.0
apispec==6.3.0
blinker==1.6.3
build==1.0.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
Flask==3.0.0
flask-apispec==0.11.4
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.2
psycopg2-binary==2.9.9
pyproject_hooks==1.0.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.22
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.6
webargs==8.3.0
Werkzeug==3.0.0
zipp==3.17.0
